+++ libs/langgraph/src/tests/pregel.test-d.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import "../graph/zod/plugin.js";
-import { z } from "zod";
+import { z } from "zod/v3";
 
-import { expectTypeOf, it, beforeAll } from "vitest";
+import { expectTypeOf, it, beforeAll, expect } from "vitest";
 import type { BaseMessage } from "@langchain/core/messages";
 import { StateGraph } from "../graph/state.js";
 import { Annotation } from "../graph/annotation.js";
@@ -10,7 +10,9 @@ import { gatherIterator } from "../utils.js";
 import { StreamMode } from "../pregel/types.js";
 import { task, entrypoint } from "../func/index.js";
 import { initializeAsyncLocalStorageSingleton } from "../setup/async_local_storage.js";
-import { INTERRUPT, isInterrupted } from "../constants.js";
+import { INTERRUPT, isInterrupted, START } from "../constants.js";
+import { withLangGraph } from "../graph/zod/meta.js";
+import { Runtime } from "../pregel/runnable_types.js";
 
 beforeAll(() => {
   // Will occur naturally if user imports from main `@langchain/langgraph` endpoint.
@@ -161,6 +163,140 @@ it("state graph annotation", async () => {
         ]
     )[]
   >();
+});
+
+it("state graph configurable", async () => {
+  const graph = new StateGraph(
+    Annotation.Root({
+      foo: Annotation({
+        reducer: (state: string[], update: string | string[]) => {
+          return Array.isArray(update)
+            ? [...state, ...update]
+            : [...state, update];
+        },
+        default: () => [],
+      }),
+    }),
+    Annotation.Root({ modelName: Annotation<string> })
+  )
+    .addNode("one", (state, config) => {
+      expectTypeOf(state).toExtend<{ foo: string[] }>();
+      expectTypeOf(config.configurable).toExtend<
+        { modelName: string } | undefined
+      >();
+
+      return { foo: "one" };
+    })
+    .addEdge(START, "one")
+    .compile();
+
+  await graph.invoke({ foo: "bar" }, { configurable: { modelName: "valid" } });
+
+  // @ts-expect-error - Invalid configurable value
+  await graph.invoke({ foo: "bar" }, { configurable: { modelName: 123 } });
+
+  const graphZod = new StateGraph(
+    z.object({
+      foo: withLangGraph(z.array(z.string()), {
+        reducer: {
+          schema: z.union([z.string(), z.array(z.string())]),
+          fn: (state: string[], update: string | string[]) => {
+            return Array.isArray(update)
+              ? [...state, ...update]
+              : [...state, update];
+          },
+        },
+      }),
+    }),
+    z.object({ modelName: z.string() })
+  )
+    .addNode("one", (state, runtime) => {
+      expectTypeOf(state).toExtend<{ foo: string[] }>();
+      expectTypeOf(runtime.configurable).toExtend<
+        { modelName: string } | undefined
+      >();
+
+      return { foo: "one" };
+    })
+    .addEdge(START, "one")
+    .compile();
+
+  await graphZod.stream(
+    { foo: "bar" },
+    { streamMode: "custom", configurable: { modelName: "valid" } }
+  );
+
+  await expect(
+    // @ts-expect-error - Invalid configurable value
+    graphZod.invoke({ foo: "bar" }, { configurable: { modelName: 123 } })
+  ).rejects.toThrow("Expected string, received number");
+});
+
+it("state graph context", async () => {
+  const graph = new StateGraph(
+    Annotation.Root({
+      foo: Annotation({
+        reducer: (state: string[], update: string | string[]) => {
+          return Array.isArray(update)
+            ? [...state, ...update]
+            : [...state, update];
+        },
+        default: () => [],
+      }),
+    }),
+    Annotation.Root({ modelName: Annotation<string> })
+  )
+    .addNode("one", (state, runtime: Runtime<{ modelName: string }>) => {
+      expectTypeOf(state).toExtend<{ foo: string[] }>();
+      expectTypeOf(runtime.context).toExtend<
+        { modelName: string } | undefined
+      >();
+      return { foo: "one" };
+    })
+    .addEdge(START, "one")
+    .compile();
+
+  await graph.invoke({ foo: "bar" }, { context: { modelName: "valid" } });
+
+  // @ts-expect-error - Invalid context value, but only checked at type-level
+  await graph.invoke({ foo: "bar" }, { context: { modelName: 123 } });
+
+  const graphZod = new StateGraph(
+    z.object({
+      foo: withLangGraph(z.array(z.string()), {
+        reducer: {
+          schema: z.union([z.string(), z.array(z.string())]),
+          fn: (state: string[], update: string | string[]) => {
+            return Array.isArray(update)
+              ? [...state, ...update]
+              : [...state, update];
+          },
+        },
+      }),
+    }),
+    z.object({ modelName: z.string() })
+  )
+    .addNode("one", (state, runtime) => {
+      expectTypeOf(state).toExtend<{ foo: string[] }>();
+      expectTypeOf(runtime.context).toExtend<
+        { modelName: string } | undefined
+      >();
+      expect(runtime.context?.modelName).toBeTypeOf("string");
+
+      return { foo: "one" };
+    })
+    .addEdge(START, "one")
+    .compile();
+
+  await graphZod.stream(
+    { foo: "bar" },
+    { streamMode: "custom", context: { modelName: "valid" } }
+  );
+
+  await expect(
+    // @ts-expect-error - Invalid context value
+    graphZod.invoke({ foo: "bar" }, { context: { modelName: 123 } })
+  ).rejects.toThrow("Expected string, received number");
 });
 
 it("state graph zod", async () => {